Stop using CG* types outside renderers
Using CGFloat/CGPoint/CGRect, and especially AffineTransform/CGAffineTransform in a cross-platform context seems inadvisable.

+++ Sources/CollectionsBenchmark/BenchmarkCLI/BenchmarkCLI+Library+Render.swift
@@ -115,7 +115,7 @@ extension _BenchmarkCLI.Library {
             let data = try renderer.render(
               graphics,
               format: format.rawValue,
-              bitmapScale: CGFloat(options.scale))
+              bitmapScale: options.scale)
             try data.write(to: url)
             count += 1
           }
